Add modules for navigation
